refactor: reduce to a command-only relay; drop editor-context
Zoom navigation (Tower -> builders -> builder -> file) is driven by the
overview API plus controller-side state and the command relay, not by an
editor feed. editor-context only gated the controller's buttons, so dropping
it loses button gray-out polish (verbs already self-guard and no-op on a
mispress), not capability.

Removes editor-context, wants-context, presence/refcount/heartbeat, and the
provider's context emitter. What remains is a stateless command relay: a
controller POSTs /api/command, Tower fans the canonical verb to the focused
provider window. Renames editor.ts -> command.ts and editor-relay.ts ->
command-relay.ts (no editor observation left to justify the name).
