--- packages/types/src/editor.ts
@@ -1,86 +1,0 @@
-/**
- * Wire contracts for Tower's editor + command channels.
- *
- * These let a CONTROLLER (an external control device or companion app) drive and
- * observe the active editor PROVIDER (the VSCode extension today, the web
- * dashboard later) over Tower's existing SSE + REST transport:
- *  - controller -> Tower: REST POST `/api/editor/*` (context demand, presence)
- *    and `/api/command` (run a canonical verb on the active provider).
- *  - provider  -> Tower: REST POST `/api/editor/context`.
- *  - Tower -> clients: SSE `editor-wants-context` / `command` to the provider;
- *    `editor-context` to controllers.
- *
- * The command channel carries CANONICAL VERBS (`view-diff`, `forward-hunk`, ...),
- * not provider-specific command ids, so one controller drives any provider; each
- * provider maps the verb to its own implementation. Pure wire shapes only.
- */
-
-/**
- * Focused-editor context that gates a controller's context verbs (forward,
- * comment). `diffFocused`: the active editor is a builder diff; `hasSelection`:
- * that diff has a non-empty selection; `artifactFocused`: the active editor is a
- * codev artifact (spec/plan/review).
- */
-export interface EditorContext {
-  diffFocused: boolean;
-  hasSelection: boolean;
-  artifactFocused: boolean;
-}
-
-// ----- SSE payloads (Tower -> client, carried as the event `body`) -----
-
-/** Tower signals whether any controller currently wants editor-context reports. */
-export interface WantsEditorContext {
-  wanted: boolean;
-}
-
-// ----- REST request/response bodies -----
-
-/**
- * Controller -> Tower (`/api/command`): run a canonical verb on the active
- * editor provider. The verb (e.g. `view-diff`, `forward-hunk`) is
- * provider-agnostic; the provider maps it to its own implementation. `args`
- * carries verb operands (typically the target builder id).
- */
-export interface CommandRequest {
-  verb: string;
-  args?: unknown[];
-}
-
-/** Provider -> Tower (`/api/editor/context`): the focused editor's context, or null. */
-export interface EditorContextReport {
-  value: EditorContext | null;
-}
-
-/** Result of a relayed command (the `/api/command` response). */
-export interface CommandResult {
-  ok: boolean;
-  error?: string;
-}
-
-// ----- Wire protocol names (single source for routes + event types) -----
-// The route paths and SSE event-type names ARE the contract: the controller,
-// Tower, and the provider must agree on them. Defining them once here (rather
-// than repeating string literals in each package) keeps the protocol in lockstep
-// and gives compile-time references instead of stringly-typed coupling.
-
-/** REST routes for the editor + command channels (client <-> Tower). */
-export const EDITOR_ROUTES = {
-  command: '/api/command',
-  context: '/api/editor/context',
-  wantsContext: '/api/editor/wants-context',
-  heartbeat: '/api/editor/heartbeat',
-} as const;
-
-/** Prefix the Tower router uses to delegate the editor routes. */
-export const EDITOR_ROUTE_PREFIX = '/api/editor/';
-
-/** SSE event-type names (carried as the envelope `type`). */
-export const EDITOR_EVENTS = {
-  /** Tower -> provider: run a canonical command verb. */
-  command: 'command',
-  /** Tower -> provider: start/stop emitting editor context. */
-  wantsContext: 'editor-wants-context',
-  /** Tower -> controllers: the focused editor's context (or null). */
-  context: 'editor-context',
-} as const;
