--- packages/vscode/src/__tests__/editor-relay.test.ts
@@ -1,180 +1,0 @@
-/**
- * Tests for the VSCode editor provider (the editor + command relay).
- *
- * Mocks `vscode` (the established pattern from overview-cache.test.ts) with a
- * controllable window so we can drive focus, the active editor, and the change
- * events, plus a fake ConnectionManager that lets us fire synthetic SSE
- * envelopes and inspect the TowerClient relay calls.
- */
-
-import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
-
-vi.mock('vscode', () => {
-  const activeListeners: Array<() => void> = [];
-  const windowListeners: Array<() => void> = [];
-  const selectionListeners: Array<() => void> = [];
-  const window = {
-    state: { focused: true },
-    activeTextEditor: undefined as unknown,
-    onDidChangeActiveTextEditor: (l: () => void) => {
-      activeListeners.push(l);
-      return { dispose: () => {} };
-    },
-    onDidChangeWindowState: (l: () => void) => {
-      windowListeners.push(l);
-      return { dispose: () => {} };
-    },
-    onDidChangeTextEditorSelection: (l: () => void) => {
-      selectionListeners.push(l);
-      return { dispose: () => {} };
-    },
-  };
-  return {
-    window,
-    commands: { executeCommand: vi.fn() },
-    Disposable: {
-      from: (...ds: Array<{ dispose?: () => void }>) => ({
-        dispose: () => ds.forEach((d) => d.dispose?.()),
-      }),
-    },
-    __control: {
-      fireActive: () => activeListeners.forEach((l) => l()),
-      fireSelection: () => selectionListeners.forEach((l) => l()),
-    },
-  };
-});
-
-// The relay imports getDiffInjectEntry to tell whether the active editor is a
-// builder diff; stub it so the test's minimal vscode mock is enough.
-vi.mock('../diff-inject-codelens.js', () => ({ getDiffInjectEntry: () => undefined }));
-
-const vscode = (await import('vscode')) as unknown as {
-  window: { state: { focused: boolean }; activeTextEditor: unknown };
-  commands: { executeCommand: ReturnType<typeof vi.fn> };
-  __control: { fireActive: () => void; fireSelection: () => void };
-};
-const { wireEditorProvider } = await import('../editor-relay.js');
-
-function makeEditor() {
-  return {
-    document: { lineCount: 800, uri: { toString: () => 'file:///x.ts', fsPath: '/x.ts' } },
-    selection: { active: { line: 5, character: 0 }, isEmpty: true },
-  };
-}
-
-function makeConnMgr(client: unknown) {
-  let sse: ((e: { type: string; data: string }) => void) | null = null;
-  return {
-    mgr: {
-      onSSEEvent: (l: (e: { type: string; data: string }) => void) => {
-        sse = l;
-        return { dispose: () => { sse = null; } };
-      },
-      getClient: () => client,
-    },
-    // Tower sends {type, title, body:JSON} on the SSE data field, no event: name.
-    fire: (type: string, payload: unknown) =>
-      sse?.({ type: '', data: JSON.stringify({ type, title: type, body: JSON.stringify(payload) }) }),
-  };
-}
-
-describe('wireEditorProvider', () => {
-  let client: { request: ReturnType<typeof vi.fn> };
-
-  // The relay POSTs via the existing TowerClient.request(path, { method, body }).
-  function requestsTo(path: string): unknown[] {
-    return client.request.mock.calls.filter((c) => c[0] === path).map((c) => JSON.parse(c[1].body));
-  }
-  function lastRequestTo(path: string): unknown {
-    const all = requestsTo(path);
-    return all[all.length - 1];
-  }
-
-  beforeEach(() => {
-    client = { request: vi.fn().mockResolvedValue({ ok: true }) };
-    vscode.window.state.focused = true;
-    vscode.window.activeTextEditor = makeEditor();
-    vscode.commands.executeCommand.mockClear();
-  });
-
-  afterEach(() => {
-    vi.restoreAllMocks();
-  });
-
-  it('reports an initial context when Tower signals wanted:true', () => {
-    const { mgr, fire } = makeConnMgr(client);
-    const disposable = wireEditorProvider(mgr as never);
-
-    fire('editor-wants-context', { wanted: true });
-
-    expect(lastRequestTo('/api/editor/context')).toEqual({
-      value: { diffFocused: false, hasSelection: false, artifactFocused: false },
-    });
-    disposable.dispose();
-  });
-
-  it('throttles context changes to one report per window', () => {
-    const nowSpy = vi.spyOn(Date, 'now');
-    nowSpy.mockReturnValue(1000);
-    const { mgr, fire } = makeConnMgr(client);
-    wireEditorProvider(mgr as never);
-
-    fire('editor-wants-context', { wanted: true }); // initial emit at t=1000
-    nowSpy.mockReturnValue(1050);
-    vscode.__control.fireSelection(); // within throttle -> dropped
-    nowSpy.mockReturnValue(1200);
-    vscode.__control.fireSelection(); // past throttle -> emitted
-
-    expect(requestsTo('/api/editor/context')).toHaveLength(2);
-  });
-
-  it('does not emit context when the window is not focused', () => {
-    vscode.window.state.focused = false;
-    const { mgr, fire } = makeConnMgr(client);
-    wireEditorProvider(mgr as never);
-
-    fire('editor-wants-context', { wanted: true });
-    expect(requestsTo('/api/editor/context')).toHaveLength(0);
-  });
-
-  it('maps a canonical verb to its VSCode command and runs it with args', async () => {
-    const { mgr, fire } = makeConnMgr(client);
-    wireEditorProvider(mgr as never);
-
-    fire('command', { verb: 'open-terminal', args: ['spir-809'] });
-    await new Promise((r) => setTimeout(r, 0));
-
-    expect(vscode.commands.executeCommand).toHaveBeenCalledWith('codev.openBuilderById', 'spir-809');
-  });
-
-  it('ignores a verb that is not in the provider map (the allowlist)', async () => {
-    const { mgr, fire } = makeConnMgr(client);
-    wireEditorProvider(mgr as never);
-
-    fire('command', { verb: 'kill-everything', args: [] });
-    await new Promise((r) => setTimeout(r, 0));
-
-    expect(vscode.commands.executeCommand).not.toHaveBeenCalled();
-  });
-
-  it('does not run a relayed verb when the window is not focused (single active provider)', async () => {
-    vscode.window.state.focused = false;
-    const { mgr, fire } = makeConnMgr(client);
-    wireEditorProvider(mgr as never);
-
-    fire('command', { verb: 'open-terminal', args: ['spir-809'] });
-    await new Promise((r) => setTimeout(r, 0));
-
-    expect(vscode.commands.executeCommand).not.toHaveBeenCalled();
-  });
-
-  it('coerces non-array verb args to an empty arg list (no crash on a stray object)', async () => {
-    const { mgr, fire } = makeConnMgr(client);
-    wireEditorProvider(mgr as never);
-
-    fire('command', { verb: 'refresh-overview', args: { not: 'an array' } });
-    await new Promise((r) => setTimeout(r, 0));
-
-    expect(vscode.commands.executeCommand).toHaveBeenCalledWith('codev.refreshOverview');
-  });
-});
